Fixed compiler warning.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@8735 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/states_screens/race_result_gui.cpp
@@ -594,7 +594,7 @@ void RaceResultGUI::displayOneEntry(unsigned int x, unsigned int y,
     // if we have highscores, move items to the left to have some more room...
     if (m_highscore_player != NULL)
     {
-        x -= m_table_width/3.0f;
+        x -= (int)(m_table_width/3.0f);
     }
     
 #ifdef USE_PER_LINE_BACKGROUND
